// packages/hitslop/src/wire/preview.generated.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PageRequest } from "./page.generated";

export type PreviewRequest = { "type": "page", id: number, request: PageRequest, } | { "type": "resource", id: number, attachmentId: string, offset: number, length: number, };